Code warning refactoring
