+++ packages/ai/src/squiggle/squiggleCodeWarnings.ts
